Updated to latest linux kernel v4.1.12 willy LTS and other improvements
Alongside with the update to the newest Long Term Support 4.1.12 kernel, the installation of Popcorn Time is removed as it is not working because of legal issues at the moment. Also upgraded the LibreOffice application to the latest version, added compatibility with Linux Mint as well as a new icon theme.